requirements.txt
# Core Libraries
transformers>=4.38.0
sentence-transformers>=2.2.2
langchain>=0.2.0
langchain-community>=0.0.22
faiss-cpu
PyPDF2
gradio>=4.0.0

# Optional (model-specific or performance)
torch>=2.0.0
accelerate
huggingface-hub

# For PDF handling & text splitting
tqdm